split fs utility
